#44 Support reading files from templates
