+++ microconfig-core/src/main/java/io/microconfig/core/properties/resolvers/expression/ExpressionEvaluator.java
@@ -15,9 +15,6 @@ class ExpressionEvaluator {
     private final EvaluationContext context;
 
     public static ExpressionEvaluator withFunctionsFrom(Class<?>... functionClasses) {
-//        EvaluationContext context = forReadOnlyDataBinding()
-//                .withInstanceMethods()
-//                .build();
         EvaluationContext context = new StandardEvaluationContext();
 
         of(functionClasses)
